Corregido error de tipo en compute_loss

+++ CustomFaster_MP_TPU/rpn.py
@@ -326,8 +326,8 @@ class RegionProposalNetwork(torch.nn.Module):
         """
 
         sampled_pos_inds, sampled_neg_inds = self.fg_bg_sampler(labels)
-        sampled_pos_inds = torch.where(torch.cat(sampled_pos_inds, dim=0))[0]
-        sampled_neg_inds = torch.where(torch.cat(sampled_neg_inds, dim=0))[0]
+        sampled_pos_inds = torch.where(torch.cat(sampled_pos_inds, dim=0))[0].to(torch.int32)
+        sampled_neg_inds = torch.where(torch.cat(sampled_neg_inds, dim=0))[0].to(torch.int32)
 
         sampled_inds = torch.cat([sampled_pos_inds, sampled_neg_inds], dim=0)
